begin refactor of execution to support for-loops
